refactor: 공모주 서비스 클래스 통합
